v9

+++ include/hitboxmenu.h
@@ -5,17 +5,18 @@
 
 #pragma once
 
+#include <common.h>
+#include <spm/dispdrv.h>
+
 #include "childmenu.h"
 #include "nandsettings.h"
-
-#include <spm/dispdrv.h>
 
 namespace mod {
 
 class HitboxMenu final : public ChildMenu
 {
 private:
-    static void activateCheck(s8 cameraId, u8 renderMode, f32 z, spm::dispdrv::DispCallback * cb, void * cbParam);
+    static void activateCheck(s32 cameraId, u8 renderMode, f32 z, spm::dispdrv::DispCallback * cb, void * cbParam);
     void optionToggle(MenuButton * button, bool &setting);
 
 public:
